+++ resources/fixtures/data/geographic-numeric-map-custom-bucket-colors.json
@@ -1,7 +1,7 @@
 {
   "title": "FIXTURE: Geographisch Deutschland (Landkreise) - Benutzerdefinierte Bucketfarben",
   "subtitle": "Subtitle",
-  "baseMap": "geographicDELandkreise",
+  "baseMap": "de-landkreise-geographic",
   "version": "2020-01-01T00:00:00.000Z",
   "entityType": "ags",
   "data": [
