+++ resources/fixtures/data/geographic-numeric-map.json
@@ -1,7 +1,7 @@
 {
   "title": "FIXTURE: Geographisch Deutschland (Landkreise) - Nummerisch",
   "subtitle": "Subtitle",
-  "baseMap": "geographicDELandkreise",
+  "baseMap": "de-landkreise-geographic",
   "version": "2020-01-01T00:00:00.000Z",
   "entityType": "ags",
   "data": [
