+++ resources/fixtures/data/hexagon-categorical-map-with-empty.json
@@ -1,7 +1,7 @@
 {
   "title": "FIXTURE: Hexagon Schweiz (Kantone) - Kategorien mit empty-Wert",
   "subtitle": "Subtitle",
-  "baseMap": "hexagonCHCantons",
+  "baseMap": "ch-cantons-hexagon",
   "version": "1997-01-01T00:00:00.000Z",
   "entityType": "name",
   "data": [
