+++ resources/fixtures/data/hexagon-convert.json
@@ -1,6 +1,6 @@
 {
   "title": "FIXTURE: Hexagon Schweiz (Kantone) - Convert big numbers",
-  "baseMap": "hexagonCHCantons",
+  "baseMap": "ch-cantons-hexagon",
   "version": "1997-01-01T00:00:00.000Z",
   "entityType": "name",
   "data": [
