+++ resources/fixtures/data/hexagon-kmeans-average-center.json
@@ -1,7 +1,7 @@
 {
   "title": "FIXTURE: Hexagon Schweiz (Kantone) - Jenks Natural Breaks Average center",
   "subtitle": "Subtitle",
-  "baseMap": "hexagonCHCantons",
+  "baseMap": "ch-cantons-hexagon",
   "version": "1997-01-01T00:00:00.000Z",
   "entityType": "name",
   "data": [
