+++ resources/fixtures/data/hexagon-kmeans.json
@@ -1,6 +1,6 @@
 {
   "acronym": "phk.",
-  "baseMap": "hexagonCHCantons",
+  "baseMap": "ch-cantons-hexagon",
   "version": "1997-01-01T00:00:00.000Z",
   "entityType": "name",
   "data": [
