+++ resources/fixtures/data/hexagon-median.json
@@ -1,6 +1,6 @@
 {
   "title": "FIXTURE: Hexagon Schweiz (Kantone) - show median",
-  "baseMap": "hexagonCHCantons",
+  "baseMap": "ch-cantons-hexagon",
   "version": "1997-01-01T00:00:00.000Z",
   "entityType": "name",
   "data": [
